ACDdbnorm {dbnorm}	R Documentation

Adjusted coefficient of determination for a data normalized for across batch signal drift

Description

This function gives a quick notification about the performance of the statistical models implemented in the dbnorm package such as Giordan (2013) and/or empirical Bayes methods in two setting of parametric and non-parametric as described by Johnson et al.(2007) and in sva package by Leek et al.(2012). It calculates and plots adjusted coefficient of determination or Adjusted R-Squared for each variable estimated in a regression model for its dependency to the batch level in the raw data and treated data via either of those models. Immediately, a score calculated based on the maximum variability explained by the batch level presents the performance of applied models. This score notifies the consistency of a model performance for all detected features (variables), facilitating quick comparison of the models for selecting one of those models, which is more appropriate to the data structure. This function is suggested for less than 2000 features (variables) to keep maximum computational speed.

Usage


ACDdbnorm(m)

Arguments

m
A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch levels in the first column.

Details

Zero and NA values are not allowed. Optionally missing value can be imputed by emvf or emvd, functions implemented in dbnorm package. Input data must be normalized prior.

Value

Several graphs compiled into a PDF file which are a correlation plot for each of applied models, a grouped barplot presenting the maximum variability associated with batch levels in the raw and the corrected datasets.
Files saved as csv in the working directory are a dataset corrected via either of applied models. Also, a two column matrix for Adjusted R-Square for raw and corrected datasets and a table summarizing the score values presented in barplot.

References

M.Giordan (2013) https://link.springer.com/article/10.1007/s12561-013-9081-1
Johnson et al. (2007) http://www.ncbi.nlm.nih.gov/pubmed/16632515
Leek et al. (2012) https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/

Examples



batch<- rep(gl(3,7,labels = c(1:3)),1)
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)

[Package dbnorm version 0.1.0 Index]