Visdbnorm {dbnorm}	R Documentation

Visualization of signal drift across batch normalization

Description

This function performs batch effect adjustment via three statistical models implemented in the dbnorm, namely two-stage procedure as described by Giordan (2013) and/or empirical Bayes methods in two setting of parametric and non-parametric as described by Johnson et al.(2007) and in sva package by Leek et al.(2012). Meanwhile, the graphical inferences in the context of unsupervised learning algorithms create visual inspection to inform users about the spatial separation of the sample sets analyzed in the different analytical runs alongside the distribution of the features (variables) in the raw and treated datasets. This function is suggested for less than 2000 features (variables) to speed up the computational process.

Usage


Visdbnorm(f)

Arguments

f
A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch in the first column.

Details

Zero and NA values are not allowed. Optionally missing value can be imputed by emvf or emvd, functions implemented in the dbnorm package. Input data must be normalized prior.

Value

Three datasets, adjusted by either of applied statistical algorithms prepared in csv and together with series of plot such as PCA plot and Scree plot compiled into a PDF file are saved in the working directory. RLA plots are represented in the Viewer panel of rstudio.

References

M.Giordan (2013) https://link.springer.com/article/10.1007/s12561-013-9081-1
Johnson et al. (2007) http://www.ncbi.nlm.nih.gov/pubmed/16632515
Leek et al. (2012) https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/

Examples



batch<- rep(gl(3,7,labels = c(1:3)),1)
y<- matrix(rnorm(2100),nrow=21)
f<-data.frame(batch,y)

[Package dbnorm version 0.1.0 Index]