dbnormPcom {dbnorm}	R Documentation

drift across batch normalization via Parametric- ComBat model and visualization base on unsupervised learning algorithm and regression analysis

Description

This function allows you to adjust the data for across batch signal drift or batch effect parametric Empirical Bayes approach as described by Johnson et al.(2007) and in sva package as explained by Leek et al.(2012). emphdbnormPcom includes advanced statistical tools to inspect the structure and quality of high throughput experiment both in macroscopic and microscopic scales at the level of sample sets and metabolic feature, respectively. Notably, using this function users applied unsupervised learning algorithm to visualize the most variance explained by the two first components in the different set of samples analyzed in the entire experiment in the raw and corrected data. In parallel, linear association of feature (variable) and batch level has been estimated and visualized by a correlation plot. In fact, estimated Adjusted- R squared is considered to define the level of dependency of feature (variable) to the batch level in the raw and corrected datasets. Besides, for quick notification about the performance of the applied model a maximum variability detected in either of datasets is reported as a score. This score notify the consistency of model performance for all detected features (variables).

Usage


dbnormPcom(m)

Arguments

m
A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch levels in the first column.

Details

Zero and NA values are not allowed. Optionally missing value can be imputed by functions implemented in dbnorm package. Input must be normalized and transformed prior.

Value

Several graphs compiled into a PDF file are a PCA score plot, Scree plot and a correlation plot for raw and corrected data. Also, the RLA plots for each dataset visualized in the Viewer panel in the rstudio console.
Files saved as csv in the working directory are a dataset corrected by the applied model. Also, a two column matrix for Adjusted R-Square raw and corrected dataset and a table summarizing the maximum score.

References

Johnson et al. (2007) http://www.ncbi.nlm.nih.gov/pubmed/16632515
Leek et al. (2012) https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/

Examples



batch<- rep(gl(3,7,labels = c(1:3)),1)
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)

[Package dbnorm version 0.1.0 Index]