emvd {dbnorm}	R Documentation

Estimation of missing value data-based

Description

It returns to a matrix of data in which missing values are estimated by the lowest detected value in the entire experiment. By this function, all NA values are replaced by Zero values, that of being ultimately replaced by the lowest value detected in the experiment. Ultimately, data matrix is transposed to restore original structure.

Usage


emvd(m)

Arguments

m
An array or a matrix

Details

empty entries are not allowed

Value

A matrix with estimated missing value.

Examples


m<- data.frame(x1=c(50,NA,6,10,30),x2=c(2,8,NA,15,0))

[Package dbnorm version 0.1.0 Index]