emvf {dbnorm}	R Documentation

Estimation of missing value feature-based

Description

It returns to a matrix of data in which missing values (Zero and/or NA values) are estimated. By this function, all Zero values are first replaced by NA values, which are then replaced by the lowest detected value on the column margin.

Usage


emvf(m)

Arguments

m
An array or a matrix

Details

empty entries are not allowed

Value

A matrix with estimated missing value.

Examples


m<- data.frame(x1=c(50,NA,6,10,30),x2=c(2,8,NA,15,0))


[Package dbnorm version 0.1.0 Index]