profplotnpcom {dbnorm}	R Documentation

Visualization of analytical heterogeneity on the profile of features (variables) in Non-Parametric ComBat corrected data

Description

Visualization of analytical heterogeneity on the profile of features (variables) in Non-Parametric ComBat corrected data

Usage


profplotnpcom(m)

Arguments

m
A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch in the first column.

Details

Zero and NA values are not allowed. Optionally missing value can be imputed by functions implemented in dbnorm package. Input must be normalized and transformed prior.

Value

Original and adjusted datasets in csv format together with the series of profile plot of the features (variables) in the sample sets provided by Scatter plot,Violin plot and pdf plot compiled into a PDF file.

References

Johnson et al. (2007) http://www.ncbi.nlm.nih.gov/pubmed/16632515
Leek et al. (2012) https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3307112/

Examples



batch<- rep(gl(3,7,labels = c(1:3),1))
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)


[Package dbnorm version 0.1.0 Index]