profplotraw {dbnorm}	R Documentation

Visualization of analytical heterogeneity on the profile of features (variables) in raw data

Description

This function informs you about the presence of across batch signal drift or batch effect in the raw data determined by the shifted probability density function plots (pdf plots) of features (variables) detected in an experiment.

Usage


profplotraw(m)

Arguments

m
A data frame in which rows define the independent experiments (samples) and columns the features (variables), with the batch level in the first column.

Details

Zero and NA values are not allowed. Optionally missing value can be imputed by functions implemented in dbnorm package. Input must be normalized and transformed prior.

Value

Original dataset in csv format together with the series of profile plot for the features (variables) in the sample sets analyzed in the entire experiment provided by the Scatter plot,Violin plot and pdf plot compiled into PDF file.

Examples



batch<- rep(gl(3,7,labels = c(1:3),1))
y<- matrix(rnorm(2100),nrow=21)
m<-data.frame(batch,y)

[Package dbnorm version 0.1.0 Index]